Fixed some time management bugs I've found playing in DOCCC.. :-O
git-svn-id: https://svn.code.sf.net/p/redqueenchess/code/redqueen@228 865e88f3-cc1b-4e5f-86d1-2a7dcd23a5c5

+++ src/searchagent.h
@@ -280,10 +280,10 @@ private:
 
 	size_t hashSize;
 	int threadNumber;
-	int whiteTime;
-	int whiteIncrement;
-	int blackTime;
-	int blackIncrement;
+	long whiteTime;
+	long whiteIncrement;
+	long blackTime;
+	long blackIncrement;
 	int depth;
 	int movesToGo;
 	int moveTime;
@@ -292,7 +292,7 @@ private:
 	size_t activeHash;
 	std::vector<TranspositionTable<Key,HashData>*> transTable;
 
-	const int getTimeToSearch();
+	const long getTimeToSearch();
 
 };
